[#61] init 관련 코드를 별도 파일로 분리
