+++ src/executor/mod.rs
@@ -4,6 +4,7 @@ pub mod encoder;
 #[allow(clippy::module_inception)]
 pub mod executor;
 pub mod implements;
+pub mod initializer;
 pub mod predule;
 pub mod reduce;
 pub mod result;
